requirements.txt
python-dotenv
openai
python-telegram-bot==20.0
pytz
timezonefinder
swisseph
geopy
pyswisseph
telegram
flask
